mobile responsive with bugs
Android browsers seem to be working well, but having trouble in Safari
for iOS and Chrome for iOS.

Safari: After clicking the ‘Start’ button, the mobile form appears to
be fixed to the bottom of the page and not sticky to the viewport.

Chrome: After clicking the ‘Start’ button, the viewport centres on the
email input and the div appears to be fixed in that spot on the page.
Viewport can scroll down or up past the footer.